+++ cinemapedia/lib/presentation/views/views.dart
@@ -1,0 +1,2 @@
+export 'movies/favorites_view.dart';
+export 'movies/home_view.dart';
